third_party/WebKit/LayoutTests/fast/css/lazy-parsing-pseudo-attribute.html

<!DOCTYPE html>
<link type="text/css" rel="stylesheet" href=".match::before { content: attr(pass); }"/>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<p>You should see the word PASS below.</p>
<div class="nomatch" pass="FAIL"></div>
<script>
test(function() {
  document.body.offsetTop;
  var elm = document.querySelector("div");
  elm.className = "match";
  document.body.offsetTop;
  elm.setAttribute("pass", "PASS");
  assert_equals(elm.getAttribute("pass"), "PASS");
}, "Test that we lazily parse content: attr(value) in ::before and ::after");
</script>
